Updated tests to run independently of user instance of app.
Tests updated so that they are running with a temporary folder as a config folder which is deleted after the tests are executed (either successfully or not). The test data are not mixed with user data of a developer.

Refers to the following issue:
https://github.com/yoda-pa/yoda/issues/116

Changes
- a function is executed prior to each test which creates a temporary folder and changes the yoda config path to this folder
- a function is executed after each test which deletes the folder and restores the original yoda config path
- therefore, all constants refering to file paths (in all modules) have been changed to functions, so that they are able to react to config path changes (otherwise it could result in a weird behaviour)
(changing file path constants to dynamically resolved functions also helps in cases where two instances of yoda are run simultaneously and one of them changes the config path)